feature: swap ic_list to ic_grid fix: base adapter, LaunchAdapter, LaunchViewModel, model launch, Mass.java, Height.java

+++ app/src/main/java/android/eservices/spacex/data/api/model/rocket/Height.java
@@ -1,7 +1,7 @@
 package android.eservices.spacex.data.api.model.rocket;
 
 public class Height {
-    private int meters;
+    private double meters;
 
-    public int getMeters() { return meters; }
+    public double getMeters() { return meters; }
 }
